Offene Frage auswerten

Gibt die Antworten auf eine offene Frage als Tabelle aus. Kommen Antworten mehrfach vor (Groß- und Kleinschreibung wird dabei ignoriert), zeigt die Tabelle jede Antwort einmal mit ihrer Häufigkeit, sortiert nach Häufigkeit.

Bei Berichten mit Anhang (appendix = TRUE, Standard) erscheint an der Stelle der Frage nur die Überschrift mit einem Link in den Anhang. Die Antworten werden gesammelt und am Ende des Berichts mit appendix_open() ausgegeben.

Usage

merge_open(
  x,
  inkl = "nr",
  inkl_global = setanalysis_defaults$inkl.open,
  nr = "",
  freq = "auto",
  appendix = setanalysis_defaults$open.appendix,
  is_appendix = FALSE,
  anchor = FALSE
)

Arguments

x

Vektor (Text) mit den Antworten. Erwartet das Attribut label (Fragetext). Fehlende Antworten (NA) werden ignoriert.

inkl

Soll die Frage im Bericht erscheinen? TRUE oder FALSE. Beim Standard "nr" entscheidet die Variable inkl.<nr> (z. B. inkl.2.1 bei nr = "2.1"), die im Berichts-Template gesetzt ist, meist aus der Berichtstabelle von input_tabelle(). Ohne Fragenummer wird die Frage immer ausgegeben.

inkl_global

Zweiter Schalter, der alle offenen Fragen eines Berichts gemeinsam ein- oder ausschließt. Voreinstellung aus setanalysis_defaults (inkl.open).

nr

Fragenummer, z. B. "2.1". Sie wird der Überschrift vorangestellt und bestimmt bei inkl = "nr", welche Variable inkl.<nr> abgefragt wird. Standard: "" (keine Nummer).

freq

Sollen gleiche Antworten zusammengefasst und mit Häufigkeit gezeigt werden? TRUE, FALSE oder "auto" (nur, wenn mindestens eine Antwort mehrfach vorkommt).

appendix

Sollen die Antworten im Anhang statt an Ort und Stelle stehen? Voreinstellung aus setanalysis_defaults (open.appendix).

is_appendix, anchor

Werden intern von appendix_open() gesetzt, um die Ausgabe im Anhang mit Rücksprung-Link zu erzeugen.

Value

Nichts (unsichtbar NULL). Der Code für den Bericht wird mit cat() ausgegeben.

See also

Examples

# Antworten direkt an Ort und Stelle
merge_open(BspDaten$dataSHOWUP$offen, appendix = FALSE)
#> ###  [FILTER_123] Welche weiteren Informationen hätten Sie gerne gehabt, damit Sie sich vor Beginn Ihres Studiums ausreichend über den Studiengang informiert gefühlt hätten? 
#>  
#> 
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | **Antwort**                                                                                                                                                                                                                                | **Häufigkeit** |
#> +============================================================================================================================================================================================================================================+================+
#> | Lorem ipsum dolor sit amet.                                                                                                                                                                                                                | 4              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Sed do eiusmod tempor.                                                                                                                                                                                                                     | 2              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Aliqua laborum anim cillum esse non consectetur nostrud. Fugiat nostrud anim laborum excepteur elit laborum.                                                                                                                               | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Cillum elit ullamco reprehenderit aliqua voluptate consequat officia enim eiusmod sed aliquip.                                                                                                                                             | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Ea nulla non esse consequat ullamco elit aute magna.                                                                                                                                                                                       | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Elit consequat aute dolore. Velit reprehenderit enim culpa sint qui ullamco ullamco esse eiusmod ut excepteur et.                                                                                                                          | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Elit lorem labore eiusmod ea nostrud culpa. Non sint occaecat aute sint.                                                                                                                                                                   | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Exercitation id excepteur non.                                                                                                                                                                                                             | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | In cupidatat dolore eiusmod nostrud ut elit aliqua.                                                                                                                                                                                        | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Incididunt quis sed sunt in. Quis voluptate officia sed dolore aliquip culpa deserunt. Pariatur occaecat nisi incididunt fugiat cupidatat minim incididunt consectetur excepteur occaecat et ipsum laborum.                                | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Laboris id sunt cillum excepteur.                                                                                                                                                                                                          | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Laborum ipsum adipiscing ea aute ad incididunt cillum enim nostrud aliqua ut velit labore. Nostrud sint officia nostrud id sit irure commodo sint fugiat. Consequat irure elit enim minim veniam eiusmod enim reprehenderit labore cillum. | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Mollit lorem et tempor. Duis nulla reprehenderit sit veniam esse ex. Reprehenderit irure cillum sunt ullamco.                                                                                                                              | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Nisi aliqua ea culpa dolore sint. Laborum nisi non sed officia nisi voluptate duis consequat nostrud. Qui consequat minim lorem dolor.                                                                                                     | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Non adipiscing in dolore ad proident ad est veniam mollit enim quis.                                                                                                                                                                       | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Nostrud ipsum cillum exercitation mollit labore ad aute. Excepteur proident ullamco dolore esse mollit excepteur sunt do. Reprehenderit in reprehenderit sint culpa elit et.                                                               | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Nulla elit culpa anim irure aute culpa do fugiat fugiat laborum excepteur dolore tempor. Nulla lorem commodo pariatur fugiat velit amet velit proident nisi exercitation consequat qui minim.                                              | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Proident in do in. Elit sit minim ut ex enim aute ex ullamco fugiat consectetur quis elit culpa.                                                                                                                                           | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Quis et minim qui commodo lorem sed sint sit exercitation id officia aliqua mollit. Aute mollit esse consectetur do officia aliqua qui in ad minim officia sit.                                                                            | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Sed id consequat velit lorem sit id duis.                                                                                                                                                                                                  | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Sint officia laborum exercitation exercitation in sed est dolor nulla proident culpa.                                                                                                                                                      | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Sit ex in sunt deserunt sunt.                                                                                                                                                                                                              | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+ 
#> 

# Bericht mit Anhang: zuerst nur Überschrift und Link, am Ende der Anhang
merge_open(BspDaten$dataSHOWUP$offen, appendix = TRUE)
#> ###  [FILTER_123] Welche weiteren Informationen hätten Sie gerne gehabt, damit Sie sich vor Beginn Ihres Studiums ausreichend über den Studiengang informiert gefühlt hätten? {#sec-2.top} 
#> 
#> *Die offenen Antworten zu dieser Frage finden sich* [im Anhang](#sec-2.bottom).  
#> 
#> \
#> 
appendix_open()
#> # Anhang: Fragen mit offenem Antwortformat  
#>   
#> ###  [FILTER_123] Welche weiteren Informationen hätten Sie gerne gehabt, damit Sie sich vor Beginn Ihres Studiums ausreichend über den Studiengang informiert gefühlt hätten? {#sec-1.bottom} 
#>  
#> [zurück nach oben](#sec-1.top) 
#> 
#> 
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | **Antwort**                                                                                                                                                                                                                                | **Häufigkeit** |
#> +============================================================================================================================================================================================================================================+================+
#> | Lorem ipsum dolor sit amet.                                                                                                                                                                                                                | 4              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Sed do eiusmod tempor.                                                                                                                                                                                                                     | 2              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Aliqua laborum anim cillum esse non consectetur nostrud. Fugiat nostrud anim laborum excepteur elit laborum.                                                                                                                               | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Cillum elit ullamco reprehenderit aliqua voluptate consequat officia enim eiusmod sed aliquip.                                                                                                                                             | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Ea nulla non esse consequat ullamco elit aute magna.                                                                                                                                                                                       | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Elit consequat aute dolore. Velit reprehenderit enim culpa sint qui ullamco ullamco esse eiusmod ut excepteur et.                                                                                                                          | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Elit lorem labore eiusmod ea nostrud culpa. Non sint occaecat aute sint.                                                                                                                                                                   | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Exercitation id excepteur non.                                                                                                                                                                                                             | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | In cupidatat dolore eiusmod nostrud ut elit aliqua.                                                                                                                                                                                        | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Incididunt quis sed sunt in. Quis voluptate officia sed dolore aliquip culpa deserunt. Pariatur occaecat nisi incididunt fugiat cupidatat minim incididunt consectetur excepteur occaecat et ipsum laborum.                                | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Laboris id sunt cillum excepteur.                                                                                                                                                                                                          | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Laborum ipsum adipiscing ea aute ad incididunt cillum enim nostrud aliqua ut velit labore. Nostrud sint officia nostrud id sit irure commodo sint fugiat. Consequat irure elit enim minim veniam eiusmod enim reprehenderit labore cillum. | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Mollit lorem et tempor. Duis nulla reprehenderit sit veniam esse ex. Reprehenderit irure cillum sunt ullamco.                                                                                                                              | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Nisi aliqua ea culpa dolore sint. Laborum nisi non sed officia nisi voluptate duis consequat nostrud. Qui consequat minim lorem dolor.                                                                                                     | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Non adipiscing in dolore ad proident ad est veniam mollit enim quis.                                                                                                                                                                       | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Nostrud ipsum cillum exercitation mollit labore ad aute. Excepteur proident ullamco dolore esse mollit excepteur sunt do. Reprehenderit in reprehenderit sint culpa elit et.                                                               | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Nulla elit culpa anim irure aute culpa do fugiat fugiat laborum excepteur dolore tempor. Nulla lorem commodo pariatur fugiat velit amet velit proident nisi exercitation consequat qui minim.                                              | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Proident in do in. Elit sit minim ut ex enim aute ex ullamco fugiat consectetur quis elit culpa.                                                                                                                                           | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Quis et minim qui commodo lorem sed sint sit exercitation id officia aliqua mollit. Aute mollit esse consectetur do officia aliqua qui in ad minim officia sit.                                                                            | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Sed id consequat velit lorem sit id duis.                                                                                                                                                                                                  | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Sint officia laborum exercitation exercitation in sed est dolor nulla proident culpa.                                                                                                                                                      | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Sit ex in sunt deserunt sunt.                                                                                                                                                                                                              | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+ 
#> 
#> ###  [FILTER_123] Welche weiteren Informationen hätten Sie gerne gehabt, damit Sie sich vor Beginn Ihres Studiums ausreichend über den Studiengang informiert gefühlt hätten? {#sec-2.bottom} 
#>  
#> [zurück nach oben](#sec-2.top) 
#> 
#> 
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | **Antwort**                                                                                                                                                                                                                                | **Häufigkeit** |
#> +============================================================================================================================================================================================================================================+================+
#> | Lorem ipsum dolor sit amet.                                                                                                                                                                                                                | 4              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Sed do eiusmod tempor.                                                                                                                                                                                                                     | 2              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Aliqua laborum anim cillum esse non consectetur nostrud. Fugiat nostrud anim laborum excepteur elit laborum.                                                                                                                               | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Cillum elit ullamco reprehenderit aliqua voluptate consequat officia enim eiusmod sed aliquip.                                                                                                                                             | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Ea nulla non esse consequat ullamco elit aute magna.                                                                                                                                                                                       | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Elit consequat aute dolore. Velit reprehenderit enim culpa sint qui ullamco ullamco esse eiusmod ut excepteur et.                                                                                                                          | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Elit lorem labore eiusmod ea nostrud culpa. Non sint occaecat aute sint.                                                                                                                                                                   | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Exercitation id excepteur non.                                                                                                                                                                                                             | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | In cupidatat dolore eiusmod nostrud ut elit aliqua.                                                                                                                                                                                        | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Incididunt quis sed sunt in. Quis voluptate officia sed dolore aliquip culpa deserunt. Pariatur occaecat nisi incididunt fugiat cupidatat minim incididunt consectetur excepteur occaecat et ipsum laborum.                                | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Laboris id sunt cillum excepteur.                                                                                                                                                                                                          | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Laborum ipsum adipiscing ea aute ad incididunt cillum enim nostrud aliqua ut velit labore. Nostrud sint officia nostrud id sit irure commodo sint fugiat. Consequat irure elit enim minim veniam eiusmod enim reprehenderit labore cillum. | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Mollit lorem et tempor. Duis nulla reprehenderit sit veniam esse ex. Reprehenderit irure cillum sunt ullamco.                                                                                                                              | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Nisi aliqua ea culpa dolore sint. Laborum nisi non sed officia nisi voluptate duis consequat nostrud. Qui consequat minim lorem dolor.                                                                                                     | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Non adipiscing in dolore ad proident ad est veniam mollit enim quis.                                                                                                                                                                       | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Nostrud ipsum cillum exercitation mollit labore ad aute. Excepteur proident ullamco dolore esse mollit excepteur sunt do. Reprehenderit in reprehenderit sint culpa elit et.                                                               | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Nulla elit culpa anim irure aute culpa do fugiat fugiat laborum excepteur dolore tempor. Nulla lorem commodo pariatur fugiat velit amet velit proident nisi exercitation consequat qui minim.                                              | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Proident in do in. Elit sit minim ut ex enim aute ex ullamco fugiat consectetur quis elit culpa.                                                                                                                                           | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Quis et minim qui commodo lorem sed sint sit exercitation id officia aliqua mollit. Aute mollit esse consectetur do officia aliqua qui in ad minim officia sit.                                                                            | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Sed id consequat velit lorem sit id duis.                                                                                                                                                                                                  | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Sint officia laborum exercitation exercitation in sed est dolor nulla proident culpa.                                                                                                                                                      | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+
#> | Sit ex in sunt deserunt sunt.                                                                                                                                                                                                              | 1              |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------+ 
#>