Rücklauf der Lehrveranstaltungen auswerten

Berechnet für jede Lehrveranstaltung den Rücklauf (Anzahl der Antworten geteilt durch die Zahl der angemeldeten Teilnehmenden, in Prozent) und gibt eine Tabelle mit Kennwerten sowie einen Boxplot (boxplot_rueck()) aus.

Da die Teilnehmendenzahl meist bei der Anmeldung zur Evaluation angegeben wird, kann der Rücklauf über 100 % liegen.

Usage

merge_rueck(x, kennung)

Arguments

x

Vektor mit der angegebenen Teilnehmendenzahl der jeweiligen Lehrveranstaltung (in jeder Zeile einer Lehrveranstaltung derselbe Wert).

kennung

Vektor mit der Kennung der Lehrveranstaltung für jede Zeile.

Value

Nichts (unsichtbar NULL). Der Code für den Bericht wird mit cat() ausgegeben.

See also

Examples

merge_rueck(BspDaten$dataLVE$Teilnehmer, BspDaten$dataLVE$Kennung)
#> 
#> +-------+-------+--------+---------+---------+
#> | **n** | **M** | **SD** | **Min** | **Max** |
#> +=======+=======+========+=========+=========+
#> | 449   | 33.9  | 44.3   | 3.1     | 600     |
#> +-------+-------+--------+---------+---------+  
#>   
#> <div class="figure" style="text-align: center">
#> <img src="figure/sub_chunk_34-1.png" alt="plot of chunk sub_chunk_34" width="100%" />
#> <p class="caption">plot of chunk sub_chunk_34</p>
#> </div>