Workload der Lehrveranstaltungen auswerten

Gibt den Fragetext als Überschrift und einen Boxplot des angegebenen Workloads (Stunden pro Woche) aus (boxplot_wl()). Standardmäßig wird zuerst der Median je Lehrveranstaltung gebildet, sodass jede Lehrveranstaltung gleich stark eingeht.

Usage

merge_wl(WL, kennung, already.aggr = FALSE)

Arguments

WL

Vektor mit dem angegebenen Workload in Stunden pro Woche. Erwartet das Attribut label (Fragetext).

kennung

Vektor mit der Kennung der Lehrveranstaltung für jede Antwort. Nicht nötig bei already.aggr = TRUE.

already.aggr

Liegt der Workload schon je Lehrveranstaltung vor? Dann wird nicht noch einmal aggregiert.

Value

Nichts (unsichtbar NULL). Der Code für den Bericht wird mit cat() ausgegeben.

See also

Examples

merge_wl(BspDaten$dataLVE$WL, kennung = BspDaten$dataLVE$Kennung)
#> ### Zusätzlich zu Ihren Anwesenheitszeiten in der Veranstaltung: Wie viel Zeit (in Zeitstunden) haben Sie für die vorliegende Veranstaltung im Schnitt pro Woche aufgewendet? (ohne Prüfungsvorbereitung) 
#> 
#> 
#> 
#> <div class="figure" style="text-align: center">
#> <img src="figure/sub_chunk_44-1.png" alt="plot of chunk sub_chunk_44" width="100%" />
#> <p class="caption">plot of chunk sub_chunk_44</p>
#> </div>