Veraltete Funktionsnamen

Diese Funktionen sind ältere Schreibweisen aktueller Funktionen. Sie verhalten sich genau wie die jeweils aktuelle Funktion und werden nur noch bereitgestellt, damit bestehende Berichtsvorlagen weiter funktionieren. In neuem Code sollten die aktuellen Namen verwendet werden.

Veralteter Name	Aktueller Name

appendix.open()	appendix_open()

boxplot.ruecklauf()	merge_rueck()

bsp.boxplot()	bsp_boxplot()

bsp.evasys.sk6()	bsp_evasys_sk6()

bsp.table.stat()	bsp_table_stat()

change.analysis.defaults()	change_analysis_defaults()

evasys.read.data()	evasys_read_data()

grade()	merge_grade()

input.tabelle()	input_tabelle()

label.test()	label_test()

list.open.answers	list_open_answers

markdown.in.viewer()	markdown_in_viewer()

merge.evasys.sk()	merge_sk()

merge.fachsem()	merge_fachsem()

merge.mc()	merge_mc()

merge.multi.sk()	merge_aggr_sk()

merge.num()	merge_num()

merge.open()	merge_open()

merge.sc()	merge_sc()

merge.subj()	merge_subj()

merge.wl()	merge_wl()

open.answers()	merge_open() mit appendix = TRUE

table.freq()	table_freq()

table.stat.multi()	table_stat_multi()

table.stat.single()	table_stat_single()

Value

Siehe die jeweils aktuelle Funktion.

Details

Die veralteten Namen leiten jeden Aufruf unverändert an die aktuelle Funktion weiter. Argumente werden dabei genau so zugeordnet wie bei einem direkten Aufruf der aktuellen Funktion.

Namen wie merge.sc() oder boxplot.ruecklauf() sehen für R wie S3-Methoden der generischen Funktionen merge() bzw. boxplot() aus. Sie haben deshalb die Argumente der generischen Funktion, damit sie nicht als fehlerhafte Methoden gelten.

list.open.answers ist ein zweiter Name für dieselbe Umgebung wie list_open_answers.